refactor: 공개 일기 확인, 일기 상세조회 test code 분리
