+++ apps/api/src/diary/spec/diary.service.spec.ts
@@ -62,125 +62,6 @@ describe('DiaryService', () => {
     uploadFileService = module.get(UploadFileService);
   });
 
-  describe('getOpenDiary', () => {
-    it('공개 일기장 확인', async () => {
-      const mockDiary: DiaryEntity = {
-        id: 1,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-        status: true,
-        isOpen: true,
-        userId: 1,
-        title: 'title',
-        content: 'content',
-      };
-      diaryRepository.findByUnique.mockResolvedValue(mockDiary);
-
-      const result = await diaryService.validateOpenDiary(1);
-
-      expect(diaryRepository.findByUnique).toHaveBeenCalledTimes(1);
-      expect(result).toBeUndefined();
-    });
-
-    it('공개된 일기장이 아닐 경우 NotFoundDiaryException 을 반환한다', async () => {
-      const mockDiary = {
-        status: true,
-        isOpen: false,
-      };
-
-      diaryRepository.findByUnique.mockResolvedValue(mockDiary);
-
-      await expect(async () => {
-        await diaryService.validateOpenDiary(1);
-      }).rejects.toThrow(new NotFoundDiaryException());
-    });
-
-    it('status 가 false 일 경우 NotFoundDiaryException 을 반환한다', async () => {
-      const mockDiary = {
-        status: false,
-        isOpen: true,
-      };
-      diaryRepository.findByUnique.mockResolvedValue(mockDiary);
-
-      await expect(async () => {
-        await diaryService.validateOpenDiary(1);
-      }).rejects.toThrow(new NotFoundDiaryException());
-    });
-
-    it('존재하지 않는 일기장일 경우 NotFoundDiaryException 을 반환한다', async () => {
-      diaryRepository.findByUnique.mockResolvedValue(null);
-
-      await expect(async () => {
-        await diaryService.validateOpenDiary(1);
-      }).rejects.toThrow(new NotFoundDiaryException());
-    });
-  });
-
-  describe('getDiary', () => {
-    it('일기 상세 조회 성공', async () => {
-      const mockDiary: DiaryIncludeImageAndLike = {
-        id: 1,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-        status: true,
-        isOpen: true,
-        userId: 1,
-        title: 'title',
-        content: 'content',
-        diaryImage: [],
-        diaryLike: [],
-      };
-      diaryRepository.findIncludeLikeAndImage.mockResolvedValue(mockDiary);
-
-      const { diaryLike, diaryImage, ...diary } = await diaryService.getDiary({
-        diaryId: 1,
-      });
-
-      expect(diaryRepository.findIncludeLikeAndImage).toHaveBeenCalledTimes(1);
-      expect(diary.status).toBe(true);
-      expect(diary.isOpen).toBe(true);
-    });
-
-    it('공개된 일기장이 아닐 경우 NotFoundDiaryException 을 반환한다', async () => {
-      const mockData = {
-        status: true,
-        isOpen: false,
-      };
-
-      diaryRepository.findIncludeLikeAndImage.mockResolvedValue(mockData);
-
-      await expect(async () => {
-        await diaryService.getDiary({
-          diaryId: 1,
-        });
-      }).rejects.toThrow(new NotFoundDiaryException());
-    });
-
-    it('status 가 false 일 경우 NotFoundDiaryException 을 반환한다', async () => {
-      const mockData = {
-        status: false,
-        isOpen: true,
-      };
-      diaryRepository.findIncludeLikeAndImage.mockResolvedValue(mockData);
-
-      await expect(async () => {
-        await diaryService.getDiary({
-          diaryId: 1,
-        });
-      }).rejects.toThrow(new NotFoundDiaryException());
-    });
-
-    it('존재하지 않는 일기장일 경우 NotFoundDiaryException 을 반환한다', async () => {
-      diaryRepository.findIncludeLikeAndImage.mockResolvedValue(null);
-
-      await expect(async () => {
-        await diaryService.getDiary({
-          diaryId: 1,
-        });
-      }).rejects.toThrow(new NotFoundDiaryException());
-    });
-  });
-
   describe('getMyDiaryList', () => {
     it('내 일기장 리스트 조회 성공', async () => {
       const mockData: WithTotal<DiaryIncludeImageAndLike[]> = [
